+++ src/player/player_cache.rs
@@ -9,7 +9,7 @@ pub struct Cache {
     pub volume: f32,
     pub shuffle: bool,
     pub playlist: String,
-    pub length: usize,
+    pub length: u32,
 }
 
 pub fn fetch_cache() -> Cache {
